backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
openpyxl==3.1.2
requests==2.31.0